Add files via upload
Added some comments to Stats.cs.
Got rid of the abstract class AbstractStat, and merged it with the Stat class, because there really wasn't a need to use an abstract class. I think the Stat class is much more readable now.
Changed the way that a Stat object's "statValue" is modified by a Pokemon's associated "Nature". It used to be that every Stat object contained a reference to a Nature object, and that Stat objects internally calculated their "statValue" fields by testing some things about the Nature object they referenced, and then multiplying the existing "statValue" by the appropriate number. Now, Stat objects all just have a "modifier" field. The logic behind computing this modifier field is now external to the Stat class, and happens inside the constructor for Pokemon.